% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-epochs.R
\name{easter_monday}
\alias{easter_monday}
\title{Easter Monday day of year}
\usage{
easter_monday(years)
}
\arguments{
\item{years}{a numeric vector of \code{years}}
}
\value{
a numeric vector
}
\description{
Determine the day of the year that the Western Easter Monday falls on.
}
