% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-methods.R
\name{eom}
\alias{eom}
\title{The end of month date}
\usage{
eom(dates)
}
\arguments{
\item{dates}{a vector of dates.}
}
\value{
a date vector with the same class as \code{dates}
}
\description{
The \code{dates} are rounded to the end of their respective months.
}
\examples{
library("lubridate")
eom(ymd(20120203, 20140203))
}
