% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-epochs.R
\name{equinox}
\alias{equinox}
\title{March and September equinox}
\usage{
equinox(years, season = "mar", tz = "UTC", want_dt = FALSE)
}
\arguments{
\item{years}{a numeric vector of \code{years}}

\item{season}{defines the Equinox sought. Can be \code{mar} (default) or
\code{sep}.}

\item{tz}{time zone for which equinox instant should be returned (default:
\code{"UTC"})}

\item{want_dt}{a flag indicating whether the returned date should be in
dynamical time (default: \code{FALSE})}
}
\value{
the date-time (UTC) of the Equinox (drop sub-minute precision)
}
\description{
Determine the date/time of March and September Equinoxes.
This implements Jean Meeus' algorithm (Astronomical Algorithms
1st Ed, 1991, Chapter 26).
}
\keyword{internal}
