% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmdates-package.r
\docType{package}
\name{fmdates}
\alias{fmdates}
\alias{fmdates-package}
\title{fmdates}
\description{
Implements common date calculations relevant for specifying the economic
nature of financial market contracts that are typically defined by
International Swap Dealer Association (ISDA) legal documentation.
}
\details{
The key classes and methods introduced by this package are documented in
\code{\link{Calendar}}, \code{\link{JointCalendar}}, \code{\link[=is_good]{is_good()}}, \code{\link[=adjust]{adjust()}}, \code{\link[=shift]{shift()}} and
\code{\link[=year_frac]{year_frac()}}.
}
