% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-methods.R
\name{is_eom}
\alias{is_eom}
\title{Checks whether dates are last day of month}
\usage{
is_eom(dates)
}
\arguments{
\item{dates}{a vector of dates.}
}
\value{
a logical vector
}
\description{
This checks whether the \code{dates} provided are the last day of a
month.
}
\examples{
library("lubridate")
is_eom(ymd(20110228)) # TRUE
is_eom(ymd(20120229)) # TRUE
}

