% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-epochs.R
\name{julian_day_to_gregorian}
\alias{julian_day_to_gregorian}
\title{Determine Gregorian date from Julian day}
\usage{
julian_day_to_gregorian(julian_day)
}
\arguments{
\item{julian_day}{a numeric vector}
}
\value{
a POSIXct vector of Gregorian dates at midnight UTC time.
}
\description{
Determine Gregorian date from Julian day
}
\keyword{internal}
