% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-methods.R
\name{nth_nday}
\alias{nth_nday}
\title{\code{n^{th} n-day}}
\usage{
nth_nday(dates, nth, nday)
}
\arguments{
\item{dates}{a vector of dates.}

\item{nth}{a numeric (possibly vector) value}

\item{nday}{a numeric (possibly vector) value representing the weekday where
Sunday is represented by the numeric value \code{1}. Case-insensitive
abbreviated three letter string representations of the day will be
transformed to such numeric values.}
}
\value{
a date vector with the same class as \code{dates}
}
\description{
Determines the \code{n^{th}} n-day after \code{dates}. You might want to
use this to determine the expiry date of futures which tend to be defined as
the second Friday of the expiry month. So in that case, \code{nth} is
\code{2} and \code{nday} is \code{6} (as Sunday has a weekday number of 1).
}
\examples{
library("lubridate")
nth_nday(ymd("20120601"), 2, 6) # 2Fri, "2012-06-08"
nth_nday(ymd("20120901"), 3, 4) # 3Wed, "2012-09-19"
nth_nday(ymd("20121201"), 1, 2) # 1Mon, "2012-12-03"
nth_nday(ymd("20121215"), 4, 5) # 4Thu, "2013-01-10"
nth_nday(ymd("20121215"), 4, "thu") # 4Thu, "2013-01-10"
}

