% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-methods.R
\name{roll_date}
\alias{roll_date}
\title{Roll date to mday}
\usage{
roll_date(dates, mday)
}
\arguments{
\item{dates}{a vector of dates.}

\item{mday}{a numeric vector representing the day of the month \code{dates}
should be rolled to.}
}
\value{
a date vector with the same class as \code{dates}
}
\description{
This function rolls dates to a given date in the month. If the mday is
before the dates provided, then the dates are roll backwards - and vice
versa. If one of the \code{dates} is the last day of the month, then the
date is left unrolled. Namely, the end of month barrier is not breached.
}
\examples{
library("lubridate")
roll_date(ymd("2012-02-12"), 15)  # "2012-02-15 UTC"
roll_date(ymd("2012-02-12"), 29)  # "2012-02-29 UTC"
roll_date(ymd("2013-02-12"), 29)  # "2011-02-28 UTC"
}
