% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-calendar-classes.R, R/date-calendar-au-classes.R, R/date-calendar-ch-classes.R, R/date-calendar-eu-classes.R, R/date-calendar-gb-classes.R, R/date-calendar-hk-classes.R, R/date-calendar-joint-classes.R, R/date-calendar-jp-classes.R, R/date-calendar-no-classes.R, R/date-calendar-nz-classes.R, R/date-calendar-us-classes.R
\docType{class}
\name{Calendar}
\alias{AUCalendar}
\alias{AUMECalendar}
\alias{AUSYCalendar}
\alias{CHZHCalendar}
\alias{Calendar}
\alias{EUTACalendar}
\alias{GBLOCalendar}
\alias{HKHKCalendar}
\alias{JPTOCalendar}
\alias{JointCalendar}
\alias{NOOSCalendar}
\alias{NZAUCalendar}
\alias{NZCalendar}
\alias{NZWECalendar}
\alias{USNYCalendar}
\title{Calendar class (R6)}
\format{A \code{\link{R6Class}} generator object}
\usage{
Calendar

AUCalendar

AUSYCalendar

AUMECalendar

CHZHCalendar

EUTACalendar

GBLOCalendar

HKHKCalendar

JointCalendar

JPTOCalendar

NOOSCalendar

NZCalendar

NZAUCalendar

NZWECalendar

USNYCalendar
}
\description{
A hierarchy of classes from which a number of methods are inherited or extended.
}
\section{Methods}{

\describe{
 \item{\code{is_good(dates)}}{Return a logical vector where good \code{dates} return
 \code{TRUE}, and vice-versa. Always returns \code{TRUE} for \code{Calendar}}
 \item{\code{adjust(dates, bdc)}}{Adjust vector of \code{dates} according to the
 business day convention \code{bdc}. Will throw an error if \code{bdc} is
 not one of "u", "f", "mf", "p", "mp" or "ms" and if it isn't a scalar.
 Otherwise, will return a vector of dates with the same class as \code{dates}.}
 \item{\code{filter(dates)}}{A thin wrapper around \code{\link{Filter}} so
 that only elements of the vector \code{dates} that are good according to
 the calendar will be returned.}
 \item{\code{shift(dates, period, bdc = "u", eom_rule = TRUE)}}{Returns a
 vector of shifted \code{dates} with the same
 date class as \code{dates}. The \code{period} should be an atomic
 \code{\link{Period-class}} in the sense that only one of its slots should be
 non-zero. It must also only be one of day, month or year.
 If one of the \code{dates} is the last business day of the month, is being shifted
 by a month or year \code{period} and \code{eom_rule} is \code{TRUE} then
 the shifted date is also the last business day of the month.}
 \item{\code{locales}}{Returns the Calendar's class name as a string which
 is used by the \code{print} method.}
}
}

\section{Active bindings}{

There is one active binding \code{tz} that returns the Olsen database
identifier as a string if no value is passed to the method. Otherwise, it
sets the Calendar's time zone attribute.
}

\section{Fields}{

\code{Calendar} has no fields.
}
\references{
\href{https://en.wikipedia.org/wiki/List_of_tz_database_time_zones}{Olsen database}
}
\keyword{data}
\keyword{datasets}

