% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmdates-package.r
\docType{package}
\name{fmdates}
\alias{fmdates}
\alias{fmdates-package}
\title{fmdates}
\description{
Implements common date calculations relevant for specifying the economic
nature of financial market contracts and typically defined in ISDA legal
documentation.
}
