% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar-methods.R
\name{is}
\alias{is}
\alias{is.Calendar}
\alias{is.JointCalendar}
\title{Calendar class checkers}
\usage{
is.Calendar(x)

is.JointCalendar(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
\code{TRUE} if \code{x} inherits from \code{Calendar} or \code{JointCalendar}
(\code{is.Calendar} and \code{is.JointCalendar} respectively) and \code{FALSE} otherwise.
}
\description{
Calendar class checkers
}
\seealso{
Other calendar methods: \code{\link{adjust}},
  \code{\link{generate_schedule}}, \code{\link{is_good}},
  \code{\link{is_valid_bdc}}, \code{\link{locale}},
  \code{\link{shift}}, \code{\link{tz}}
}
